requirements.txt
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-milvus==0.1.6
langchain-text-splitters==0.3.2
langsmith==0.1.140
pymilvus==2.4.9
pdfminer==20191125
pdfminer.six==20240706
sentence-transformers==3.2.1
numpy==1.26.4
regex==2024.11.6
Flask==3.0.3
tqdm==4.67.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
umap-learn==0.5.7